+++ .gitignore
@@ -12,7 +12,6 @@
 node_modules
 .next
 out
-.env
 
 # Swap the comments on the following lines if you don't wish to use zero-installs
 # Documentation here: https://yarnpkg.com/features/zero-installs
