+++ apps/ind4/.env
@@ -1,0 +1,7 @@
+NEXT_PUBLIC_BECKN_API_URL=https://api.osm-dev.becknprotocol.io
+NEXT_PUBLIC_NOMINATIM_URL=https://nominatim.openstreetmap.org
+# login/signup authentication 
+NEXT_PUBLIC_STRAPI_URL=https://strapi-bap.becknprotocol.io/api
+#client layer url
+NEXT_PUBLIC_API_URL=http://localhost:3001/assembly
+# NEXT_PUBLIC_API_URL=https://de83-194-95-60-104.ngrok-free.app
