+++ apps/ind4/components/productDetails/DetailsSection.tsx
@@ -2,10 +2,10 @@ import { Box, Flex, Image, Text, Button } from '@chakra-ui/react'
 import React, { useEffect, useState } from 'react'
 import StarRatingComponent from 'react-star-rating-component'
 import { useLanguage } from '../../hooks/useLanguage'
+import useRequest from '../../hooks/useRequest'
+import Router from 'next/router'
+
 import { RetailItem } from '../../lib/types/products'
-import CallToAction from './CallToAction'
-import greenVegIcon from '../../public/images/greenVeg.svg'
-import redNonVegIcon from '../../public/images/redNonVeg.svg'
 
 interface Props {
   product: RetailItem
@@ -13,7 +13,97 @@ interface Props {
 const DetailsSection: React.FC<Props> = ({ product }) => {
   const { t } = useLanguage()
   const [showComponent, setShowComponent] = useState(false)
+  const apiUrl = process.env.NEXT_PUBLIC_API_URL
+  const { data, loading, error, fetchData } = useRequest()
+  const [showContentA, setShowContentA] = useState(true)
 
+  const id = product.id
+  const providerItemid = product.items[0].id
+  // "66b7b9bad166-4a3f-ada6-ca063dc9d321"
+  const providerFulfillmentid = JSON.stringify(product.items[0].fulfillment_id)
+  const providerTagname = 'select-1'
+  const bppId = 'beckn-sandbox-bpp.becknprotocol.io'
+  // const bppId = "beckn-sandbox-bpp.becknprotocol.io"
+  const bppUri = 'https://sandbox-bpp-network.becknprotocol.io/'
+  // console.log("dank item",itemId)
+
+  // const handleBook = () => {
+  // localStorage.setItem('proDet', JSON.stringify({ id: id}))
+
+  // Router.push(`/orderDetails?id=${id}`)
+
+  const selectPayload = {
+    context: {
+      bppId: bppId,
+      bppUri: bppUri
+    },
+    providerId: id,
+    itemId: providerItemid,
+    fulfillmentId: providerFulfillmentid,
+    tagName: providerTagname
+  }
+
+  const fetchDataForSelect = () => fetchData(`${apiUrl}/select`, 'POST', selectPayload)
+
+  useEffect(() => {
+    // console.log(data)
+    // const formUrl = 'http://localhost:8080'
+
+    if (data) {
+      dispatch(responseDataActions.addTransactionId(data.context.transaction_id))
+      let selectedItem = data.selectProviders.map((provider: any) => {
+        return {
+          order: {
+            provider: {
+              id: provider.id
+            },
+            descriptor: {
+              name: provider.name,
+              long_desc: provider.long_desc,
+              images: provider.image
+            },
+            formUrl: items[0].xinput.form.url
+          }
+        }
+      })
+      localStorage.setItem('selectItems', JSON.stringify(selectedItem))
+      setItems(selectedItem)
+    }
+    // window.addEventListener('beforeunload', fetchDataForSelect)
+    // return () => {
+    //   // Cleanup the event listener when the component is unmounted
+    //   window.removeEventListener('beforeunload', fetchDataForSelect)
+    // }
+    // if(data){
+    //   const formUrl = 'http://localhost:8080'
+    //   Router.push(`/orderDetails?url=${formUrl}`)
+    // }
+  }, [data])
+  const toggleContent = () => {
+    setShowContentA(!showContentA)
+  }
+  const urlNew = () => {
+    fetchDataForSelect()
+
+    const formUrl = 'http://localhost:8080'
+    Router.push(`/orderDetails?url=${formUrl}`)
+  }
+
+  const ContentA = () => (
+    <div>
+      <h2>This is Content A</h2>
+      <p>Some content for A...</p>
+    </div>
+  )
+
+  const ContentB = () => (
+    <div>
+      <h2>This is Content B</h2>
+      <p>Some content for B...</p>
+    </div>
+  )
+
+  //----
   useEffect(() => {
     localStorage.removeItem('optionTags')
     localStorage.setItem('optionTags', JSON.stringify({ name: product.descriptor.name }))
@@ -48,14 +138,6 @@ const DetailsSection: React.FC<Props> = ({ product }) => {
         <Text mt={'10px'} mb={'10px'} fontSize={'14px'}>
           {product.descriptor.short_desc}
         </Text>
-
-        {/* {product.tags.foodType ? (
-          product.tags.foodType === 'veg' ? (
-            <Image pt={'4px'} src={greenVegIcon} alt="greenVegIcon" />
-          ) : (
-            <Image pt={'4px'} src={redNonVegIcon} alt="redNonVegIcon" />
-          )
-        ) : null} */}
       </Flex>
       <hr className="mt-1 hidden md:block" />
       <div className="flex items-start flex-wrap relative ">
@@ -81,7 +163,14 @@ const DetailsSection: React.FC<Props> = ({ product }) => {
         </div>
       </div>
       <div className="flex flex-col">
-        <Button background={'rgba(var(--color-primary))'} color={'rgba(var(--text-color))'} isDisabled={false}>
+        <Button
+          background={'rgba(var(--color-primary))'}
+          color={'rgba(var(--text-color))'}
+          isDisabled={false}
+          onClick={urlNew}
+
+          // onChange={urlNew}
+        >
           Book now
         </Button>
       </div>
